2FA SMS LOGIN DONE
WHAT A SHITSHOW OF A CODEBASE TO DO THIS IN.

Now, if the user wishes to login, they must use 2FA if available. This has not been tested on those without 2FA, but Chat says that it should work.

Now, I have to start modifying the code so that it actually looks aesthetic! (Per Mias standards...)

Things to fix:
-Make it look aesthetic.
-Once the user logs in, they are AUOMATICALLY sent an SMS. The button to generate a new one will be blocked for 30 seconds with a countdown, at which they will have the option to request a new code once again.

I am so happy that this shit worked, you don't even know!!!!!!!!!!